+++ lib/skylight/messages/batch.rb
@@ -5,6 +5,7 @@ module Skylight
 
       required :timestamp, :uint32,  1
       repeated :endpoints, Endpoint, 2
+      optional :hostname,  :string,  3
 
     end
   end