+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add the ability to configure logging from railtie
 * [BUG] Handle AS::N monkey patching when there are already subscribers
+* Tracks the current host
 
 ## 0.1.6 (June 11, 2013)